+++ controllers/PostsController.php
@@ -146,7 +146,7 @@ class PostsController extends Controller
 
     /**
      * Updates an existing Blog post model.
-     * If update is successful, the browser will be redirected to the 'view' page.
+     * If update is successful, the browser will be redirected to the 'index' page.
      * @param integer $id
      * @return mixed
      * @throws NotFoundHttpException if the model cannot be found