+++ models/CategoriesSearch.php
@@ -45,6 +45,11 @@ class CategoriesSearch extends Categories
         // add conditions that should always apply here
         $dataProvider = new ActiveDataProvider([
             'query' => $query,
+            'sort'=> [
+                'defaultOrder' => [
+                    'id' => SORT_DESC
+                ]
+            ]
         ]);
 
         $this->load($params);