+++ views/posts/_form.php
@@ -127,8 +127,8 @@ $(document).ready(function() {
                 }
             ).done(function(data) {
                 if (data.alias && form.find('#blog-alias').val().length == 0) {
-                    form.find('#blog-alias').val(data.alias);
-                    form.yiiActiveForm('validateAttribute', 'blog-alias');
+                    form.find('#posts-alias').val(data.alias);
+                    form.yiiActiveForm('validateAttribute', 'posts-alias');
                 }
             }).fail(function () {
                 /*form.find('#options-type').val("");