+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 Changelog
 =========
 
+## 1.1.2 (2020-03-22)
+ * Added default sort and fix breadcrumbs
+
 ## 1.1.1 (2020-03-07)
  * Rename models, added default category, refactoring migrations
  * Filter posts by categories and tags (view dashboard)